requirements.txt
paho-mqtt
requests
opencv-python
python-openstackclient
openstacksdk
tensorflow
pandas 
seaborn 
matplotlib 
scikit-learn
arduino-iot-client
typing_extensions
edgeimpulse
ngcsdk
python-dotenv
pyyaml